src/requirements.txt
python-decouple
pandasai
streamlit
openai
requests
json
